requirements-test.txt
py-moneyed
mock
pytest-mock
pytest
pytest-django


# Additional test requirements go here
